Removed input handling-method, moved it into update().
After reading the Documentation of Godot about input handling I did remove the old inputUpdate() method as well as processInput() in the state machine to accommadate for the new way of doing it. Now I am only accessing the global Input class to see whether a key is pressed and hold down or for other inputs as well. I also got the player's paddle moving now and it collides with the ball using the internal physics enging of Godot.