Switched the behaviour for a stuck ball.
Made it so the ball gets pushed into a direction depending on it's previous direction. Furthermore I replaced the _process() method with the _integrated_forces() method to keep the ball going at a constant speed, otherwise it would loose speed over time. This was done according to the Documentation that the linear_velocity attribute is only to be touched inside this method and not inside _process() to avoid unexpected behaviour, or at least to lower the risk of it happening. Since so I also changed the same inside the randomizeStartDirection() method and used apply_central_impulse() to it instead to push it into its direction.